vader#window#execute: use a single tempfile
This not only avoids multiple calls to `tempname()`, but makes Vim
use the same script ID also when not run in Docker.
For some reason it uses a new script ID (`<SID>`/`<SNR>`) when run
normally, but uses the same ID when run in Docker.
This means that the `!` with `function! s:vader_wrapper()` was not
required in an environment that behaves like running it in Docker (with
Neomake's tests).

This also removes the `delete()`, which is handled by Vim itself when
shutting down.